Add code reformat

+++ src/main/java/me/benny/practice/spring/security/note/NoteService.java
@@ -1,6 +1,5 @@
 package me.benny.practice.spring.security.note;
 
-import java.util.List;
 import lombok.RequiredArgsConstructor;
 import me.benny.practice.spring.security.user.User;
 import me.benny.practice.spring.security.user.UserNotFoundException;
@@ -8,6 +7,8 @@ import org.springframework.data.domain.Sort;
 import org.springframework.data.domain.Sort.Direction;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+
+import java.util.List;
 
 @Service
 @Transactional
@@ -20,6 +21,7 @@ public class NoteService {
      * 노트 조회
      * 유저는 본인의 노트만 조회할 수 있다.
      * 어드민은 모든 노트를 조회할 수 있다.
+     *
      * @param user 노트를 찾을 유저
      * @return 유저가 조회할 수 있는 모든 노트 List
      */
@@ -36,8 +38,9 @@ public class NoteService {
 
     /**
      * 노트 저장
-     * @param user 노트 저장하는 유저
-     * @param title 제목
+     *
+     * @param user    노트 저장하는 유저
+     * @param title   제목
      * @param content 내용
      * @return 저장된 노트
      */
@@ -50,7 +53,8 @@ public class NoteService {
 
     /**
      * 노트 삭제
-     * @param user 삭제하려는 노트의 유저
+     *
+     * @param user   삭제하려는 노트의 유저
      * @param noteId 노트 ID
      */
     public void deleteNote(User user, Long noteId) {
@@ -58,6 +62,8 @@ public class NoteService {
             throw new UserNotFoundException();
         }
         Note note = noteRepository.findByIdAndUser(noteId, user);
-        noteRepository.delete(note);
+        if (note != null) {
+            noteRepository.delete(note);
+        }
     }
 }
